requirements.txt
# ── Core ML Stack ────────────────────────────────────────────────────────────
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# ── VLM / Transformers ────────────────────────────────────────────────────────
transformers==4.45.2
accelerate==0.30.1
peft==0.10.0
bitsandbytes==0.43.1
trl==0.8.6
qwen-vl-utils==0.0.8
einops==0.8.0
timm==1.0.3

# ── Data Pipeline ─────────────────────────────────────────────────────────────
webdataset==0.2.86
decord==0.6.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
openpack-toolkit==0.9.1

# ── API / Serving ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
httpx==0.27.0

# ── Experiment Tracking ───────────────────────────────────────────────────────
wandb==0.17.0
tensorboard==2.16.2

# ── Utilities ─────────────────────────────────────────────────────────────────
PyYAML==6.0.1
tqdm==4.66.4
rich==13.7.1
loguru==0.7.2
psutil==5.9.8
gputil==1.4.0
ffmpeg-python==0.2.0
